+++ packages/extension-polkagate/src/fullscreen/accountDetails/components/AccountIconsFs.tsx
@@ -4,20 +4,18 @@
 /* eslint-disable react/jsx-max-props-per-line */
 
 import type { DeriveAccountInfo } from '@polkadot/api-derive/types';
-import type { Option, u128, Vec } from '@polkadot/types';
 //@ts-ignore
 import type { PalletProxyAnnouncement, PalletRecoveryActiveRecovery } from '@polkadot/types/lookup';
-import type { Proxy } from '../../../util/types';
 
 import { faChain, faCheckCircle, faCircleInfo, faShieldHalved, faSitemap } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Grid, IconButton, useTheme } from '@mui/material';
-import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useContext, useEffect, useState } from 'react';
 
 import { ActionContext, Infotip } from '../../../components';
-import { useAnimateOnce, useInfo, useTranslation } from '../../../hooks';
+import { useAnimateOnce, useHasIdentityTooltipText, useHasProxyTooltipText, useInfo, useIsRecoverableTooltipText } from '../../../hooks';
 import { windowOpen } from '../../../messaging';
-import { IDENTITY_CHAINS, PROXY_CHAINS, SOCIAL_RECOVERY_CHAINS } from '../../../util/constants';
+import { IDENTITY_CHAINS } from '../../../util/constants';
 
 interface AddressDetailsProps {
   address: string | undefined;
@@ -25,71 +23,23 @@ interface AddressDetailsProps {
 }
 
 function AccountIconsFs ({ accountInfo, address }: AddressDetailsProps): React.ReactElement {
-  const { t } = useTranslation();
   const theme = useTheme();
 
   const onAction = useContext(ActionContext);
   const { account, api, chain, formatted } = useInfo(address);
 
   const [hasID, setHasID] = useState<boolean | undefined>();
-  const [isRecoverable, setIsRecoverable] = useState<boolean | undefined>();
-  const [hasProxy, setHasProxy] = useState<boolean | undefined>();
+
+  const { isRecoverable, recoverableToolTipTxt } = useIsRecoverableTooltipText(address);
+  const { hasProxy, proxyTooltipTxt } = useHasProxyTooltipText(address);
+  const identityToolTipTxt = useHasIdentityTooltipText(address, hasID);
 
   const shakeProxy = useAnimateOnce(hasProxy);
-  const shakeIdentity = useAnimateOnce(hasID);
   const shakeShield = useAnimateOnce(isRecoverable);
-
-  const identityToolTipTxt = useMemo(() => {
-    if (!chain) {
-      return 'Account is in Any Chain mode';
-    }
-
-    switch (hasID) {
-      case true:
-        return 'Has Identity';
-      case false:
-        return 'No Identity';
-      default:
-        return 'Checking';
-    }
-  }, [chain, hasID]);
-
-  const recoverableToolTipTxt = useMemo(() => {
-    if (!chain) {
-      return 'Account is in Any Chain mode';
-    }
-
-    switch (isRecoverable) {
-      case true:
-        return 'Recoverable';
-      case false:
-        return 'Not Recoverable';
-      default:
-        return 'Checking';
-    }
-  }, [chain, isRecoverable]);
-
-  const proxyTooltipTxt = useMemo(() => {
-    if (!chain) {
-      return 'Account is in Any Chain mode';
-    }
-
-    switch (hasProxy) {
-      case true:
-        return 'Has Proxy';
-      case false:
-        return 'No Proxy';
-      default:
-        return 'Checking';
-    }
-  }, [chain, hasProxy]);
+  const shakeIdentity = useAnimateOnce(hasID);
 
   useEffect((): void => {
-    setHasID(undefined);
-    setIsRecoverable(undefined);
-    setHasProxy(undefined);
-
-    if (!api || !address || !account?.genesisHash || api.genesisHash.toHex() !== account.genesisHash) {
+    if (!api || !formatted || !account?.genesisHash || api.genesisHash.toHex() !== account.genesisHash) {
       return;
     }
 
@@ -98,28 +48,7 @@ function AccountIconsFs ({ accountInfo, address }: AddressDetailsProps): React.R
     } else {
       setHasID(false);
     }
-
-    if (api.query?.['recovery'] && SOCIAL_RECOVERY_CHAINS.includes(account.genesisHash)) {
-      api.query['recovery']['recoverable'](formatted)
-        .then((r) =>
-          setIsRecoverable((r as Option<PalletRecoveryActiveRecovery>).isSome))
-        .catch(console.error);
-    } else {
-      setIsRecoverable(false);
-    }
-
-    if (api.query?.['proxy'] && PROXY_CHAINS.includes(account.genesisHash)) {
-      api.query['proxy']['proxies'](formatted)
-        .then((p) => {
-          const _p = p as unknown as [Vec<PalletProxyAnnouncement>, u128];
-          const fetchedProxies = JSON.parse(JSON.stringify(_p[0])) as unknown as Proxy[];
-
-          setHasProxy(fetchedProxies.length > 0);
-        }).catch(console.error);
-    } else {
-      setHasProxy(false);
-    }
-  }, [api, address, formatted, account?.genesisHash, accountInfo]);
+  }, [api, formatted, account?.genesisHash, accountInfo]);
 
   const openIdentity = useCallback(() => {
     address && chain && windowOpen(`/manageIdentity/${address}`).catch(console.error);
@@ -136,7 +65,7 @@ function AccountIconsFs ({ accountInfo, address }: AddressDetailsProps): React.R
   return (
     <Grid alignItems='center' container direction='column' display='grid' height='72px' item justifyContent='center' justifyItems='center' width='fit-content'>
       <Grid item onClick={openIdentity} sx={{ cursor: 'pointer', height: '24px', m: 'auto', p: '2px', width: 'fit-content' }}>
-        <Infotip placement='right' text={t(identityToolTipTxt)}>
+        <Infotip placement='right' text={identityToolTipTxt}>
           {hasID
             ? accountInfo?.identity?.displayParent
               ? <FontAwesomeIcon
@@ -157,7 +86,7 @@ function AccountIconsFs ({ accountInfo, address }: AddressDetailsProps): React.R
         </Infotip>
       </Grid>
       <Grid height='24px' item my='1px' width='24px'>
-        <Infotip placement='right' text={t(recoverableToolTipTxt)}>
+        <Infotip placement='right' text={recoverableToolTipTxt}>
           <IconButton
             onClick={openSocialRecovery}
             sx={{ height: '24px', width: '24px' }}
@@ -171,7 +100,7 @@ function AccountIconsFs ({ accountInfo, address }: AddressDetailsProps): React.R
         </Infotip>
       </Grid>
       <Grid height='24px' item width='fit-content'>
-        <Infotip placement='right' text={t(proxyTooltipTxt)}>
+        <Infotip placement='right' text={proxyTooltipTxt}>
           <IconButton onClick={openManageProxy} sx={{ height: '16px', width: '16px' }}>
             <FontAwesomeIcon
               icon={faSitemap}