+++ packages/extension-polkagate/src/hooks/useBalances.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type React from 'react';
-import type { BN } from '@polkadot/util';
 import type { BalancesInfo, SavedBalances } from '../util/types';
 
 import { useEffect, useState } from 'react';
@@ -10,10 +9,9 @@ import { useEffect, useState } from 'react';
 import { updateMeta } from '../messaging';
 import { NATIVE_TOKEN_ASSET_ID, NATIVE_TOKEN_ASSET_ID_ON_ASSETHUB } from '../util/constants';
 import { isUpToDate } from './useAssetsBalances';
-import { useBalancesOnAssethub, useBalancesOnMultiAssetChain, useInfo, useNativeAssetBalances, usePoolBalances, useStakingAccount } from '.';
+import { useBalancesOnAssethub, useBalancesOnMultiAssetChain, useInfo, useNativeAssetBalances, usePoolBalances } from '.';
 
 export default function useBalances (address: string | undefined, refresh?: boolean, setRefresh?: React.Dispatch<React.SetStateAction<boolean>>, onlyNew = false, assetId?: string | number): BalancesInfo | undefined {
-  const stakingAccount = useStakingAccount(address);
   const { account, api, chainName, decimal: currentDecimal, genesisHash: chainGenesisHash, token: currentToken } = useInfo(address);
 
   const isNativeAssetId = String(assetId) === String(NATIVE_TOKEN_ASSET_ID) || String(assetId) === String(NATIVE_TOKEN_ASSET_ID_ON_ASSETHUB);
@@ -22,7 +20,7 @@ export default function useBalances (address: string | undefined, refresh?: bool
   const balances = useNativeAssetBalances(address, refresh, setRefresh, onlyNew);
   const maybeBalancesOnAssetHub = useBalancesOnAssethub(address, maybeNonNativeAssetId);
   const maybeBalancesOnMultiChainAssets = useBalancesOnMultiAssetChain(address, maybeNonNativeAssetId);
-  const pooledBalance = usePoolBalances(address, refresh);
+  const pooledBalance = usePoolBalances(address, refresh); // can move it inside useNativeAssetBalances hook and then remove overall state var
 
   const [overall, setOverall] = useState<BalancesInfo | undefined>();
 
@@ -35,20 +33,19 @@ export default function useBalances (address: string | undefined, refresh?: bool
     if (balances && isUpToDate(balances?.date) && pooledBalance && apiGenesisHash === chainGenesisHash && apiGenesisHash === balances?.genesisHash && apiGenesisHash === pooledBalance.genesisHash) {
       setOverall({
         ...balances,
-        pooledBalance: pooledBalance.balance,
-        soloTotal: stakingAccount?.stakingLedger?.total as unknown as BN
+        pooledBalance: pooledBalance.balance
       });
     } else {
       setOverall(undefined);
     }
-  }, [pooledBalance, balances, apiGenesisHash, chainGenesisHash, stakingAccount]);
+  }, [pooledBalance, balances, apiGenesisHash, chainGenesisHash]);
 
   useEffect(() => {
     if (!address || !apiGenesisHash || apiGenesisHash !== account?.genesisHash || !overall || !chainName || !token || !decimal || account?.genesisHash !== chainGenesisHash || account?.genesisHash !== overall.genesisHash) {
       return;
     }
 
-    // TODO: this just saves native assets in local storage! can save other assets as well
+    // TODO: this just SAVES native assets in local storage! can save other assets as well
     /** to SAVE fetched balance in local storage, first load saved balances of different chaines if any */
     const savedBalances = JSON.parse(account?.balances ?? '{}') as SavedBalances;
 
@@ -81,7 +78,7 @@ export default function useBalances (address: string | undefined, refresh?: bool
 
   return overall && overall.genesisHash === chainGenesisHash && overall.token === currentToken && overall.decimal === currentDecimal
     ? overall
-    : balances && balances.token === currentToken && balances.decimal === currentDecimal
+    : balances && balances.genesisHash === chainGenesisHash && balances.token === currentToken && balances.decimal === currentDecimal
       ? balances
       : undefined;
 }