+++ packages/extension-polkagate/src/popup/home/news.ts
@@ -10,9 +10,13 @@ export interface News {
 
 export const news: News[] = [
   {
-    version: '0.18.5',
+    version: '0.21.1',
     notes: [
-      'Show No Internet Sign: Display a disconnected icon if the extension cannot access the internet.'
+      'Show No Internet Sign: Display an icon to indicate when there’s no internet connection.',
+      'Identicon Theme Selection: Allow users to choose a theme for their account identicons.',
+      'Auto-Loading Referenda: Automatically load more referenda as you scroll, removing the need for manual clicks.',
+      'Real-Time Referenda Search: Apply keyword-based search and filtering instantly, even while referenda are still loading.',
+      'Currency Conversion Display: Show asset values in the selected currency on the governance page.'
     ]
   },
   {